requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
frozenlist==1.2.0
fs==2.4.14
future==0.18.2
gcp-storage-emulator==2021.12.2
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-logging==2.7.0
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
line-bot-sdk==2.0.1
MarkupSafe==2.0.1
multidict==5.2.0
packaging==21.3
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
